# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: ระบบทำนายอัตราเลือดชิดสำหรับปศุสัตว์
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: animals
    description: Animal records
  - name: inbreeding
    description: Inbreeding coefficient calculations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /animals:
    get:
      operationId: listAnimals
      tags: [animals]
      summary: List all animals
      parameters:
        - name: farm
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of animals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Animal"
    post:
      operationId: createAnimal
      tags: [animals]
      summary: Create a new animal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnimalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Animal"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /animals/farms:
    get:
      operationId: listFarms
      tags: [animals]
      summary: List all unique farm names
      responses:
        "200":
          description: List of farm names
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  /animals/{id}:
    get:
      operationId: getAnimal
      tags: [animals]
      summary: Get an animal by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Animal details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Animal"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateAnimal
      tags: [animals]
      summary: Update an animal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnimalUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Animal"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteAnimal
      tags: [animals]
      summary: Delete an animal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /animals/{id}/pedigree:
    get:
      operationId: getAnimalPedigree
      tags: [animals]
      summary: Get pedigree tree for an animal (up to 4 generations)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pedigree tree
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PedigreeNode"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /inbreeding/calculate:
    post:
      operationId: calculateInbreeding
      tags: [inbreeding]
      summary: Calculate inbreeding coefficient (F) for a potential mating pair using A-matrix
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InbreedingRequest"
      responses:
        "200":
          description: Inbreeding calculation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InbreedingResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /inbreeding/history:
    get:
      operationId: listInbreedingHistory
      tags: [inbreeding]
      summary: List recent inbreeding calculation history
      responses:
        "200":
          description: List of past calculations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InbreedingHistory"

  /inbreeding/stats:
    get:
      operationId: getInbreedingStats
      tags: [inbreeding]
      summary: Get summary statistics
      responses:
        "200":
          description: Inbreeding statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InbreedingStats"

  /inbreeding/a-matrix:
    post:
      operationId: computeAMatrix
      tags: [inbreeding]
      summary: Compute full Additive Relationship Matrix for a set of animals
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AMatrixRequest"
      responses:
        "200":
          description: A-matrix result with F and R values
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AMatrixResult"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    Animal:
      type: object
      required: [id, name, code, species, sex, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        species:
          type: string
        sex:
          type: string
          enum: [male, female]
        farm:
          type: ["string", "null"]
        birthDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        sireId:
          type: ["integer", "null"]
        damId:
          type: ["integer", "null"]
        sireName:
          type: ["string", "null"]
        damName:
          type: ["string", "null"]
        fCoefficient:
          type: ["number", "null"]
          description: Animal's own inbreeding coefficient
        createdAt:
          type: string

    AnimalInput:
      type: object
      required: [name, code, species, sex]
      properties:
        name:
          type: string
          minLength: 1
        code:
          type: string
          minLength: 1
        species:
          type: string
          minLength: 1
        sex:
          type: string
          enum: [male, female]
        farm:
          type: string
        birthDate:
          type: string
        notes:
          type: string
        sireId:
          type: ["integer", "null"]
        damId:
          type: ["integer", "null"]

    AnimalUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        code:
          type: string
        species:
          type: string
        sex:
          type: string
          enum: [male, female]
        farm:
          type: string
        birthDate:
          type: string
        notes:
          type: string
        sireId:
          type: ["integer", "null"]
        damId:
          type: ["integer", "null"]

    PedigreeNode:
      type: object
      required: [id, name, code, sex]
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        sex:
          type: string
        fCoefficient:
          type: ["number", "null"]
        sire:
          $ref: "#/components/schemas/PedigreeNode"
        dam:
          $ref: "#/components/schemas/PedigreeNode"

    InbreedingRequest:
      type: object
      required: [sireId, damId]
      properties:
        sireId:
          type: integer
        damId:
          type: integer

    InbreedingResult:
      type: object
      required: [sireId, damId, sireName, damName, fCoefficient, rCoefficient, fSire, fDam, riskLevel, commonAncestors, pathways]
      properties:
        sireId:
          type: integer
        damId:
          type: integer
        sireName:
          type: string
        damName:
          type: string
        fCoefficient:
          type: number
          description: Predicted inbreeding coefficient of offspring F = 0.5 * A[sire][dam]
        fPercent:
          type: number
        rCoefficient:
          type: number
          description: Relationship coefficient R between sire and dam
        rPercent:
          type: number
        fSire:
          type: number
          description: Inbreeding coefficient of the sire itself
        fSirePercent:
          type: number
        fDam:
          type: number
          description: Inbreeding coefficient of the dam itself
        fDamPercent:
          type: number
        riskLevel:
          type: string
          enum: [safe, low, moderate, high, very_high]
        riskLabel:
          type: string
        commonAncestors:
          type: array
          items:
            $ref: "#/components/schemas/CommonAncestor"
        pathways:
          type: array
          items:
            type: string

    CommonAncestor:
      type: object
      required: [id, name, code, contribution]
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        contribution:
          type: number

    InbreedingHistory:
      type: object
      required: [id, sireId, damId, sireName, damName, fCoefficient, riskLevel, calculatedAt]
      properties:
        id:
          type: integer
        sireId:
          type: integer
        damId:
          type: integer
        sireName:
          type: string
        damName:
          type: string
        fCoefficient:
          type: number
        fPercent:
          type: number
        rCoefficient:
          type: ["number", "null"]
        rPercent:
          type: ["number", "null"]
        fSire:
          type: ["number", "null"]
        fDam:
          type: ["number", "null"]
        riskLevel:
          type: string
        riskLabel:
          type: string
        calculatedAt:
          type: string

    InbreedingStats:
      type: object
      required: [totalAnimals, totalCalculations, averageF, riskBreakdown]
      properties:
        totalAnimals:
          type: integer
        totalCalculations:
          type: integer
        averageF:
          type: number
        safePairings:
          type: integer
        riskyPairings:
          type: integer
        riskBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/RiskBreakdownItem"

    RiskBreakdownItem:
      type: object
      required: [level, label, count]
      properties:
        level:
          type: string
        label:
          type: string
        count:
          type: integer

    AMatrixRequest:
      type: object
      properties:
        farm:
          type: string
          description: Filter by farm; if omitted, uses all animals

    AMatrixResult:
      type: object
      required: [animals]
      properties:
        animals:
          type: array
          items:
            $ref: "#/components/schemas/AMatrixAnimal"

    AMatrixAnimal:
      type: object
      required: [id, name, code, sex, fCoefficient, fPercent]
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        sex:
          type: string
        farm:
          type: ["string", "null"]
        fCoefficient:
          type: number
        fPercent:
          type: number

    ImportResult:
      type: object
      required: [inserted, skipped, errors]
      properties:
        inserted:
          type: integer
        skipped:
          type: integer
        errors:
          type: array
          items:
            type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string